% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(
  NOAArawData = "./data_raw/earthquakes-2021-10-06_09-24-38_+0200.tsv"
)
}
\arguments{
\item{NOAArawData}{raw data input}
}
\value{
df object with the cleaned data
}
\description{
This function cleans data previously downloaded from
'https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/event-data'.
The raw downloaded data have been stored in in data-raw
The cleaned data contains: (1) filter out of NA years, (2) Date as a composition of Year-Mo-Dy
}
\examples{
\dontrun{
df <- eq_clean_data(NOAArawData='./data_raw/earthquakes-2021-10-06_09-24-38_+0200.tsv')
}
}
