% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{just use this function in the dplyr pipeline to add the poup}
}
\value{
new popup of points
}
\description{
This functions add a new popup to the map created with eq_map
using html notation to show clean and organize data in points
}
\details{
eq_create_label
}
\examples{
\dontrun{
dfs \%>\%
mutate(popup = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup")
}
}
