% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{eq_map}
\alias{eq_map}
\title{function to create a map with the selected eq points}
\usage{
eq_map(df_map = NULL, annot_col)
}
\arguments{
\item{df_map}{df_map}

\item{annot_col}{annot_col}
}
\value{
leaflet map of a dataframe where eq are represented
}
\description{
This functions creates a map based lon longitud and latitud and
sizes the points based in Mag
}
\details{
eq_map
}
\examples{
\dontrun{
p3 <- dfs \%>\% eq_map(annot_col = "Date")
p3
}

}
