% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{eqloc}
\alias{eqloc}
\title{eqloc}
\format{
An object of class \code{eqloc} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
eqloc
}
\arguments{
\item{Geom}{geom for ggplot}

\item{required_aes}{required_aes}

\item{default_aes}{default_aes}

\item{draw_panel}{draw_panel (default)}
}
\value{
default function for ggplot representation (lines with locations)
}
\description{

}
\keyword{datasets}
